Add files via upload
This a mega project called the TinDog Project which includes the usage all of the bootstrap components, examples, icons and many more. This project and all of the other projects of the bootstrap section has taken lot of hardwork, time, consistency and patience of me. This project was a total fun and 100% interesting.